Anagh QR/QR_Scanner/requirements.txt
kivy
pyzbar
pillow